- Added annotation support - Removed some redundant types - Moved scopes to CodeModel and removed Linker - Added const members and expression support for them - Added statement transformation (can be used by annotations)

+++ CodeModel/src/main/java/org/openzen/zenscript/codemodel/expression/ConstantByteExpression.java
@@ -13,9 +13,9 @@ import org.openzen.zenscript.shared.CodePosition;
  * @author Hoofdgebruiker
  */
 public class ConstantByteExpression extends Expression {
-	public final byte value;
+	public final int value;
 	
-	public ConstantByteExpression(CodePosition position, byte value) {
+	public ConstantByteExpression(CodePosition position, int value) {
 		super(position, BasicTypeID.BYTE, null);
 		
 		this.value = value;
