+++ CodeModel/src/main/java/org/openzen/zenscript/codemodel/expression/ConstantUShortExpression.java
@@ -13,9 +13,9 @@ import org.openzen.zenscript.shared.CodePosition;
  * @author Hoofdgebruiker
  */
 public class ConstantUShortExpression extends Expression {
-	public final short value;
+	public final int value;
 	
-	public ConstantUShortExpression(CodePosition position, short value) {
+	public ConstantUShortExpression(CodePosition position, int value) {
 		super(position, BasicTypeID.USHORT, null);
 		
 		this.value = value;
